Move root components into src
