+++ src/components/FeaturedCategories.tsx
@@ -1,6 +1,6 @@
 // components/FeaturedCategories.tsx
 import Image from 'next/image';
-import styles from '../styles/FeaturedCategories.module.css';
+import styles from '../../styles/FeaturedCategories.module.css';
 
 type Category = {
   id: string;
